refactor: vertical logo 및 dialog 파일 위치 수정
