+++ src/assets/svgs/VerticalLogo.tsx
